# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ChatListEvent:
    properties:
      event_start:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_host:
        type: boolean
      title:
        type: string
    type: object
...
host: staging-api.spotlightinfo.in
info:
  contact: {}
  description: REST API for Spotlight — a community events platform
  title: Spotlight API
  version: "1.0"
paths:
...
  /auth/me:
    delete:
      description: Anonymizes the currently authenticated user's account and removes
        associated data
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to anonymize user" or "Failed to delete registrations"
            or "Failed to commit
          schema:
            type: string
      summary: Delete current user account
      tags:
      - auth
    get:
      description: Retrieves the profile information of the currently authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: Returns a JSON object containing the user's profile information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Error occurred while selecting user details
          schema:
            type: string
      summary: Get current user profile
      tags:
      - auth
    patch:
      consumes:
      - multipart/form-data
      description: Updates the profile information of the currently authenticated
        user
      parameters:
      - description: Name
        in: formData
        name: name
        type: string
      - description: Phone
        in: formData
        name: phone
        type: string
      - description: Bio
        in: formData
        name: bio
        type: string
      - description: City
        in: formData
        name: city
        type: string
      - description: Date of Birth (YYYY-MM-DD)
        in: formData
        name: date_of_birth
        type: string
      - description: Gender (male, female, non_binary, prefer_not_to_say)
        in: formData
        name: gender
        type: string
      - description: Age (13-120)
        in: formData
        name: age
        type: integer
      - description: Onboarding Complete
        in: formData
        name: onboarding
        type: boolean
      - description: Avatar Image
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid form data" or "Invalid gender value" or "Age must be
            between 13 and 120
          schema:
            type: string
        "500":
          description: Failed to update profile" or "Failed to read avatar" or "Failed
            to upload avatar
          schema:
            type: string
      summary: Update current user profile
      tags:
      - auth
  /auth/request-otp:
    post:
      consumes:
      - application/json
      description: Sends an OTP to the provided email address for authentication
      parameters:
      - description: Request body containing the email address
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid json" or "Email required
          schema:
            type: string
        "500":
          description: Error in hashing" or "Inserting OTP failed" or "Failed to send
            OTP email
          schema:
            type: string
      summary: Request OTP for email login
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verifies the provided OTP for the given email address and returns
        a JWT token if successful
      parameters:
      - description: Request body containing the email address and OTP
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: JWT token and onboarding status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body" or "Invalid OTP
          schema:
            type: string
        "500":
          description: Failed to create user" or "Failed to create token
          schema:
            type: string
      summary: Verify OTP for email login
      tags:
      - auth
  /chats:
    get:
      description: Returns a list of events that the currently authenticated user
        is either hosting or attending. Each event includes its ID, title, image URL,
        start time, and whether the user is the host.
      produces:
      - application/json
      responses:
        "200":
          description: Returns a JSON object containing a list of chat events
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handlers.ChatListEvent'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: 'Internal Server Error: failed to fetch chats'
          schema:
            type: string
      summary: Get a list of chats for the authenticated user
      tags:
      - chat
  /chats/{id}/token:
    get:
      description: Returns a Firebase custom token for the currently authenticated
        user to access the chat for a specific event. The user must be either the
        host or an accepted attendee of the event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns a JSON object containing the chat token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid event ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: 'Forbidden: user is not an accepted attendee or host'
          schema:
            type: string
        "500":
          description: 'Internal Server Error: failed to create chat token'
          schema:
            type: string
      summary: Get a chat token for an event
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"